Rename licenseseat and related.
