+++ app/Http/Controllers/Users/UsersController.php
@@ -13,7 +13,7 @@ use App\Models\Company;
 use App\Models\Group;
 use App\Models\Ldap;
 use App\Models\LicenseModel;
-use App\Models\LicenseSeat;
+use App\Models\License;
 use App\Models\Location;
 use App\Models\Setting;
 use App\Models\User;
