+++ app/Http/Transformers/LicensesTransformer.php
@@ -1,6 +1,7 @@
 <?php
 namespace App\Http\Transformers;
 
+use App\Models\License;
 use App\Models\LicenseModel;
 use Gate;
 use Illuminate\Database\Eloquent\Collection;
@@ -9,39 +10,37 @@ use App\Helpers\Helper;
 class LicensesTransformer
 {
 
-    public function transformLicenses (Collection $licenses, $total)
+    public function transformLicenseSeats (Collection $seats, $total)
     {
         $array = array();
-        foreach ($licenses as $license) {
-            $array[] = self::transformLicense($license);
+        $seat_count = 0;
+        foreach ($seats as $seat) {
+            $seat_count++;
+            $array[] = self::transformLicenseSeat($seat, $seat_count);
         }
         return (new DatatablesTransformer)->transformDatatables($array, $total);
     }
 
-    public function transformLicense (LicenseModel $license)
+    public function transformLicenseSeat (License $seat, $seat_count)
     {
         $array = [
-            'id' => (int) $license->id,
-            'name' => e($license->name),
-            'company' => ($license->company) ? ['id' => (int) $license->company->id,'name'=> e($license->company->name)] : null,
-            'manufacturer' =>  ($license->manufacturer) ? ['id' => (int) $license->manufacturer->id,'name'=> e($license->manufacturer->name)] : null,
-            'product_key' => (Gate::allows('viewKeys', LicenseModel::class)) ? e($license->serial) : '------------',
-            'order_number' => e($license->order_number),
-            'purchase_order' => e($license->purchase_order),
-            'purchase_date' => Helper::getFormattedDateObject($license->purchase_date, 'date'),
-            'purchase_cost' => e($license->purchase_cost),
-            'notes' => e($license->notes),
-            'expiration_date' => Helper::getFormattedDateObject($license->expiration_date, 'date'),
-            'seats' => (int) $license->seats,
-            'free_seats_count' => (int) $license->free_seats_count,
-            'license_name' =>  e($license->license_name),
-            'license_email' => e($license->license_email),
-            'maintained' => ($license->maintained == 1) ? true : false,
-            'supplier' =>  ($license->supplier) ? ['id' => (int)  $license->supplier->id,'name'=> e($license->supplier->name)] : null,
-            'category' =>  ($license->category) ? ['id' => (int)  $license->category->id,'name'=> e($license->category->name)] : null,
-            'created_at' => Helper::getFormattedDateObject($license->created_at, 'datetime'),
-            'updated_at' => Helper::getFormattedDateObject($license->updated_at, 'datetime'),
-            'user_can_checkout' => (bool) ($license->free_seats_count > 0),
+            'id' => (int) $seat->id,
+            'license_id' => (int) $seat->license->id,
+            'name' => 'Seat '.$seat_count,
+            'assigned_user' => ($seat->user) ? [
+                'id' => (int) $seat->user->id,
+                'name'=> e($seat->user->present()->fullName)
+            ] : null,
+            'assigned_asset' => ($seat->asset) ? [
+                'id' => (int) $seat->asset->id,
+                'name'=> e($seat->asset->present()->fullName)
+            ] : null,
+            'location' => ($seat->location()) ? [
+                'id' => (int) $seat->location()->id,
+                'name'=> e($seat->location()->name)
+            ] : null,
+            'reassignable' => (bool) $seat->license->reassignable,
+            'user_can_checkout' => (($seat->assigned_to=='') && ($seat->asset_id=='')) ? true : false,
         ];
 
         $permissions_array['available_actions'] = [
@@ -57,9 +56,7 @@ class LicensesTransformer
         return $array;
     }
 
-    public function transformAssetsDatatable($licenses) {
-        return (new DatatablesTransformer)->transformDatatables($licenses);
-    }
+
