+++ app/Models/Asset.php
@@ -421,7 +421,7 @@ class Asset extends Depreciable
 
     public function licenseseats()
     {
-        return $this->hasMany('\App\Models\LicenseSeat', 'asset_id');
+        return $this->hasMany('\App\Models\License', 'asset_id');
     }
 
     public function supplier()
