+++ app/Models/Loggable.php
@@ -95,7 +95,7 @@ trait Loggable
     private function determineLogItemType($log)
     {
         // We need to special case licenses because of license_seat vs license.  So much for clean polymorphism :
-        if (static::class == LicenseSeat::class) {
+        if (static::class == License::class) {
             $log->item_type = LicenseModel::class;
             $log->item_id = $this->license_id;
         } else {
@@ -117,7 +117,7 @@ trait Loggable
         $log->target_type = get_class($target);
         $log->target_id = $target->id;
 
-        if (static::class == LicenseSeat::class) {
+        if (static::class == License::class) {
             $log->item_type = LicenseModel::class;
             $log->item_id = $this->license_id;
         } else {
@@ -176,7 +176,7 @@ trait Loggable
     {
         $log = new Actionlog;
         $location = Location::find($location_id);
-        if (static::class == LicenseSeat::class) {
+        if (static::class == License::class) {
             $log->item_type = LicenseModel::class;
             $log->item_id = $this->license_id;
         } else {
@@ -215,7 +215,7 @@ trait Loggable
             $user_id = Auth::user()->id;
         }
         $log = new Actionlog;
-        if (static::class == LicenseSeat::class) {
+        if (static::class == License::class) {
             $log->item_type = LicenseModel::class;
             $log->item_id = $this->license_id;
         } else {
@@ -238,7 +238,7 @@ trait Loggable
     public function logUpload($filename, $note)
     {
         $log = new Actionlog;
-        if (static::class == LicenseSeat::class) {
+        if (static::class == License::class) {
             $log->item_type = LicenseModel::class;
             $log->item_id = $this->license_id;
         } else {
