+++ database/seeds/LicenseSeeder.php
@@ -1,14 +1,14 @@
 <?php
 use Illuminate\Database\Seeder;
 use App\Models\LicenseModel;
-use App\Models\LicenseSeat;
+use App\Models\License;
 
 class LicenseSeeder extends Seeder
 {
     public function run()
     {
         LicenseModel::truncate();
-        LicenseSeat::truncate();
+        License::truncate();
         factory(LicenseModel::class, 1)->states('photoshop')->create();
         factory(LicenseModel::class, 1)->states('acrobat')->create();
         factory(LicenseModel::class, 1)->states('indesign')->create();
