+++ resources/views/companies/view.blade.php
@@ -100,7 +100,7 @@
                         <div class="table-responsive">
 
                             <table
-                                    data-columns="{{ \App\Presenters\LicensePresenter::dataTableLayout() }}"
+                                    data-columns="{{ \App\Presenters\LicenseModelPresenter::dataTableLayout() }}"
                                     data-cookie-id-table="licensesTable"
                                     data-pagination="true"
                                     data-id-table="licensesTable"
